extracting generate() function and exposing it as main index.js
