+++ index.js
@@ -1,0 +1,1 @@
+module.exports.generate = require("./lib/generate");
